refactor: repo cleanup — remove dead code, split large files, add tests (closes #225)
- Remove dead fortressDwarves module (unused outside its own test)
- Extract tile-glyphs.ts from MainViewport.tsx (deduplicate glyph maps)
- Split task-execution.ts into deprivation.ts and task-completion.ts
- Extract useDesignation and useWorldState hooks from App.tsx
- Create shared test-helpers.ts for sim test factories
- Add tests for deprivation, event-firing, task-completion, task-helpers
- Rename embark.test.ts → dwarf-names.test.ts
- Add code health rules to CLAUDE.md
